Mostly working metrics and graphs in topology
