--- ui-v2/app/components/service-sparklines/index.hbs
@@ -1,13 +1,0 @@
-<DataSource
-  @src={{uri nspace dc 'metrics' 'summary-for-service' @service 'tcp'}}
-  @onchange={{action 'change'}} />
-
-
-<div class="sparkline-wrapper" style="position: relative; margin-top: 70px;">
-  <div class="tooltip" style="visibility: hidden; position: absolute; background: #eee; padding: 10px;">
-    <span class="sparkline-time"></span>
-  </div>
-  <div class="sparkline-loader">Loading...</div>
-  <svg class="sparkline" width="300" height="100" stroke-width="2"></svg>
-</div>
-
