+++ ui-v2/app/components/topology-metrics/index.hbs
@@ -11,14 +11,15 @@
         </Tooltip>
       </span>
     </div>
-    <TopologyMetrics::Card @items={{@downstreams}} @dc={{@dc}} />
+    <TopologyMetrics::Card @items={{@downstreams}} @service={{@service.Service.Service}} @dc={{@dc}} />
   </div>
 {{/if}}
   <div id="metrics-container">
     <div>
       {{@service.Service.Service}}
     </div>
-    <TopologyMetrics::Stats @endpoint='summary-for-service' @item={{@service.Service.Service}} @protocol=''/>
+    <TopologyMetrics::Series @service={{@service.Service.Service}} @protocol='http' />
+    <TopologyMetrics::Stats @endpoint='summary-for-service' @service={{@service.Service.Service}} @protocol='http'/>
     <div class="link">
      {{#if @metricsHref}}
       <a class="metrics-link" href={{@metricsHref}} target="_blank" rel="noopener noreferrer">Open metrics Dashboard</a>
@@ -41,7 +42,7 @@
   {{#each-in (group-by "Datacenter" @upstreams) as |dc upstreams|}}
     <div id="upstream-container">
       <p>{{dc}}</p>
-      <TopologyMetrics::Card @items={{upstreams}} @dc={{@dc}} @type='upstream' />
+      <TopologyMetrics::Card @items={{upstreams}} @service={{@service.Service.Service}} @dc={{@dc}} @type='upstream' />
     </div>
   {{/each-in}}
   </div>
