+++ ui-v2/app/components/topology-metrics/layout.scss
@@ -106,11 +106,19 @@
 // Stats Section
 .stats {
   padding: 12px;
-  display: inline-flex;
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: space-between;
+  align-items: stretch;
   dl {
-    margin-right: 8px;
+    display:flex;
+  }
+  dt {
+    margin-right: 5px;
+    line-height: 1.5em !important;
   }
   dd {
     color: $gray-400 !important;
+    margin-right: 10px;
   }
 }
