+++ ui-v2/app/components/topology-metrics/series/index.scss
@@ -1,0 +1,2 @@
+@import './skin';
+@import './layout';
