+++ ui-v2/app/components/topology-metrics/series/layout.scss
@@ -1,0 +1,27 @@
+#metrics-container div .sparkline-wrapper {
+  position: relative;
+  width: 100%;
+  padding: 0;
+  margin: 0;
+  height: 70px;
+
+  svg.sparkline {
+    width: 100%;
+    height: 70px;
+    padding: 0;
+    margin: 0;
+  }
+
+  .tooltip {
+    visibility: hidden;
+    position: absolute;
+    z-index: 100;
+    bottom: 78px;
+    width: 250px;
+  }
+
+  .sparkline-tt-legend-color {
+    display: inline-block;
+  }
+}
+
