--- ui-v2/app/components/topology-metrics/stats.hbs
@@ -1,39 +1,0 @@
-<DataSource
-  @src={{uri nspace dc 'metrics' @endpoint @item @protocol}}
-  @onchange={{action (mut metrics) value="data"}}
-/>
-
-<div class="stats">
-  <dl>
-    <dt>
-      {{!-- {{metrics.stats.RPS}} --}}
-    </dt>
-    <dd>
-      RPS
-    </dd>
-  </dl>
-  <dl>
-    <dt>
-      {{!-- {{metrics.stats.ER}} --}}
-    </dt>
-    <dd>
-      ER
-    </dd>
-  </dl>
-  <dl>
-    <dt>
-      {{!-- {{metrics.stats.MED}} --}}
-    </dt>
-    <dd>
-      MED
-    </dd>
-  </dl>
-    <dl>
-    <dt>
-      {{!-- {{metrics.stats.P99}} --}}
-    </dt>
-    <dd>
-      P99
-    </dd>
-  </dl>
-</div>
