+++ ui-v2/app/services/repository/metrics.js
@@ -1,5 +1,6 @@
 import RepositoryService from 'consul-ui/services/repository';
 import { inject as service } from '@ember/service';
+import { env } from 'consul-ui/env';
 
 // meta is used by DataSource to configure polling. The cursor is fake since we
 // aren't really doing blocking queries. The pollInterval controls how long
@@ -7,7 +8,7 @@ import { inject as service } from '@ember/service';
 // TODO - make this configurable in the UI settings.
 const meta = {
   cursor: 1,
-  pollInterval: 10000,
+  pollInterval: env('CONSUL_METRICS_POLL_INTERVAL', 10000),
 }
 
 export default RepositoryService.extend({
