+++ ui-v2/app/styles/components.scss
@@ -68,3 +68,4 @@
 @import 'consul-ui/components/consul-intention-permission-header-list';
 @import 'consul-ui/components/role-selector';
 @import 'consul-ui/components/topology-metrics';
+@import 'consul-ui/components/topology-metrics/series';
