+++ ui-v2/package.json
@@ -150,5 +150,7 @@
       "lib/startup"
     ]
   },
-  "dependencies": {}
+  "dependencies": {
+    "dayjs": "^1.9.1"
+  }
 }
